requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
groq
pymongo==3.12